Game logic now works due to edits in playGame function. Added code to show winner.
the playGame function was not working as expected, now due to moving the order of items in the function it now
works as expected.
I also created some logic to display the winner of the game after 5 rounds, but this is not displayed at the
end yet. Will need to fix.